backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyPDF2
pydantic
python-dotenv
groq
gunicorn
openai
